ofGraphics: remove some warnings

+++ libs/openFrameworks/graphics/ofGraphics.cpp
@@ -1021,11 +1021,11 @@ void ofDrawBitmapStringHighlight(string text, const ofPoint& position, const ofC
 void ofDrawBitmapStringHighlight(string text, int x, int y, const ofColor& background, const ofColor& foreground) {
 	vector<string> lines = ofSplitString(text, "\n");
 	int maxLineLength = 0;
-	for(int i = 0; i < lines.size(); i++) {
+	for(int i = 0; i < (int)lines.size(); i++) {
 		// tabs are not rendered
 		const string & line(lines[i]);
 		int currentLineLength = 0;
-		for(int j = 0; j < line.size(); j++) {
+		for(int j = 0; j < (int)line.size(); j++) {
 			if (line[j] == '\t') {
 				currentLineLength += 8 - (currentLineLength % 8);
 			} else {
